+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [Unreleased]
 ### Added
 - `GlimpseAdapter.glLineWidth(float)` to set line width
+- Blending-related methods in `GlimpseAdapter`:
+  - `glEnableBlending`
+  - `glDisableBlending`
+  - `glBlendingFunction`
+  - `glBlendingColor`
 
 ### Changed
 - Make `GLCapabilitiesFactory` internal in `glimpse-ui`